[mod/#223]mapper 재정립
